Register puntuations

+++ backend/jugaripunt/migrations/0006_jugador_puntuacio.py
@@ -13,6 +13,6 @@ class Migration(migrations.Migration):
         migrations.AddField(
             model_name='jugador',
             name='puntuacio',
-            field=models.IntegerField(default=0),
+            field=models.FloatField(default=0.0),
         ),
     ]
