+++ backend/jugaripunt/models.py
@@ -12,7 +12,7 @@ class Jugador(models.Model):
     session_token = models.CharField(max_length=255, blank=True, null=True)  # Camp opcional per al token de sessió
     admin = models.BooleanField(default=False)  # per controlar si es o no administrador
     num_federat = models.IntegerField(default=0)  # numero de federat del jugador (numeric)
-    puntuacio = models.IntegerField(default=0)  # Puntuació del jugador (sencer)
+    puntuacio = models.FloatField(default=0.0)  # Puntuació del jugador (decimal)
 
     def __str__(self):
         return f"{self.nom} - {self.cognoms}"  # Retorna el nom del jugador com a representació en cadena
